config.py moved to utils
