+++ agents/termination_agent.py
@@ -3,7 +3,7 @@ import textwrap
 from autogen_agentchat.agents import AssistantAgent
 from autogen_core.models import ChatCompletionClient
 
-from config import Config
+from utils.config import Config
 
 
 class TerminationAgent(AssistantAgent):
