+++ scenario_engine/scenario_task.py
@@ -1,11 +1,11 @@
 import datetime
 import os
 
-from config import Config
 from monitoring.console_monitor_ansi import ConsoleMonitorAnsi
 from monitoring.monitor import MonitorBase
 from scenarios.orchestrator_agent_base import OrchestratorAgentBase, OrchestratorContext
 from scenarios.scenario_base import ScenarioBase
+from utils.config import Config
 from utils.misc import generate_timestamp, to_os_path
 from utils.tee import Tee
