+++ scenarios/new_app/new_app_orchestrator_agent.py
@@ -11,9 +11,9 @@ from agents.output_agent import OutputAgent
 from agents.reviewer_agent import ReviewerAgent
 from agents.scaffold_agent import ScaffoldAgent
 from agents.test_agent import TestAgent
-from config import Config
 from constants import BUILD_AGENT_SUCCESSFUL, SCAFFOLD_AGENT_DONE
 from scenarios.orchestrator_agent_base import OrchestratorAgentBase, OrchestratorContext
+from utils.config import Config
 
 
 class NewAppOrchestratorAgent(OrchestratorAgentBase):
