+++ tests/test_build_agent.py
@@ -5,8 +5,8 @@ from autogen_agentchat.messages import TextMessage
 from autogen_agentchat.ui import Console
 
 from agents.build_agent import BuildAgent
-from config import Config
 from tests.test_utils import setup_test
+from utils.config import Config
 from utils.misc import to_os_path
