+++ tests/test_fix_build_scenario.py
@@ -2,9 +2,9 @@ import os
 
 import pytest
 
-from config import Config
 from scenarios.fix_build.fix_build_scenario import FixBuildScenario
 from tests.test_utils import setup_test
+from utils.config import Config
 from utils.misc import to_os_path
